ads1256_pio.py: Change the termination sync for buffered read.
Since the SDATAC is now sent by the data state machine, the timing
at the end of data acquisition is less critical. Use a single scheduled
IRQ now, which can do the sign correction and wait for the data
state machine to finish.

Fix initialisation of the data_acquired flag.

Signed-off-by: robert-hh <robert@hammelrath.com>