ads1256: Provide a sensor_detect() method.
This small method sets or resets the sensor detect current source at
the active input. The sensor detect current sources allow to verify the
integrity of an external sensor supplying a signal to the ADS1255/6.
A shorted sensor produces a very small signal while an open-circuit
sensor produces a very large signal.

Signed-off-by: robert-hh <robert@hammelrath.com>